Move descent criterion into individual files
